refactor: fix marker test package and classname
